+++ src/http/oauth2.ts
@@ -108,29 +108,53 @@ export function createOAuth2HttpClient(
     tokenData = data;
   }
 
-  async function authenticate(data: OAuth2Config): Promise<void> {
+  async function authenticate(data: OAuth2Config): Promise<TokenDataOauth2> {
     authConfig = data;
-    const tokenResult = await http.post(options.path, {
-      ...options.params,
-      ...authConfig.params,
-    });
+    const tokenResult = await http.post(
+      options.path,
+      {
+        ...options.params,
+        ...authConfig.params,
+      },
+      options.params.client_secret
+        ? {
+            auth: {
+              username: options.params.client_id,
+              password: options.params.client_secret,
+            },
+          }
+        : {}
+    );
     setTokenData(tokenResult.data);
+    return tokenResult.data;
   }
 
   async function confirmMfa({
     token,
     methodId,
     code,
-  }: MfaConfig): Promise<void> {
-    const tokenResult = await http.post(options.path, {
-      ...options.params,
-      ...authConfig.params,
-      grant_type: 'mfa',
-      token,
-      code,
-      method_id: methodId,
-    });
+  }: MfaConfig): Promise<TokenDataOauth2> {
+    const tokenResult = await http.post(
+      options.path,
+      {
+        ...options.params,
+        ...authConfig.params,
+        grant_type: 'mfa',
+        token,
+        code,
+        method_id: methodId,
+      },
+      options.params.client_secret
+        ? {
+            auth: {
+              username: options.params.client_id,
+              password: options.params.client_secret,
+            },
+          }
+        : {}
+    );
     setTokenData(tokenResult.data);
+    return tokenResult.data;
   }
 
   function logout(): boolean {