+++ src/types.ts
@@ -36,7 +36,6 @@ export interface ParamsOauth1WithToken {
 
 export interface ParamsOauth2AuthorizationCode {
   code: string;
-  redirectUri: string;
 }
 
 export interface ParamsOauth2Password {
@@ -74,6 +73,7 @@ export interface ParamsOauth1 extends ParamsBase {
 
 export interface ParamsOauth2 extends ParamsBase {
   clientId: string;
+  clientSecret?: string;
 }
 
 interface HttpClientBase {
@@ -90,6 +90,8 @@ export interface ConfigOauth2 extends ParamsBase {
   params: {
     // eslint-disable-next-line camelcase
     client_id: string;
+    // eslint-disable-next-line camelcase
+    client_secret?: string;
   };
 }